requirements.txt
azure-cognitiveservices-speech==1.37.0
python-dotenv
sounddevice
numpy
